概要欄装飾を分割
